- fixed comments and moved helpers to core/utils/env.ts
